requirements.txt
absl-py==2.1.0
mediapipe==0.8.11
opencv-python-headless
matplotlib==3.8.3
numpy==1.26.4
streamlit==1.33.0
streamlit-option-menu==0.3.12
scikit-learn==1.4.1.post1
tensorflow==2.16.1
mysql-connector-python==8.4.0
pandas==2.0.3

